requirements.txt
asgiref==3.9.1
Django==5.2.4
django-ckeditor==6.7.3
django-extensions==4.1        # يمكن حذفه في الإنتاج
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-taggit==6.1.0
MarkupSafe==3.0.2
pillow==10.3.0
psycopg2-binary==2.9.10       # أو psycopg2
python-decouple==3.8
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
gunicorn==21.2.0              # تم تحديد الإصدار
python-dotenv==1.0.1
